# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_api.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_api.muteConvReq:
    properties:
      convId:
        type: string
    required:
    - convId
    type: object
...
  internal_api.updateSettingsReq:
    properties:
      msg_preview_enabled:
        type: boolean
      mute_list:
        type: string
      notification_enabled:
        type: boolean
    type: object
  internal_api.uploadAvatarResponse:
    properties:
      file_path:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: GoIM 即时通讯系统 API — 支持私聊、群聊、朋友圈动态、AI 助手。
  title: GoIM API
  version: "1.0"
paths:
...
  /settings:
    get:
      description: 获取当前用户的通知偏好和免打扰设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户设置
      tags:
      - 设置
    put:
      consumes:
      - application/json
      description: 更新当前用户的通知偏好和免打扰设置
      parameters:
      - description: 设置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.updateSettingsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 更新用户设置
      tags:
      - 设置
  /settings/mute:
    post:
      consumes:
      - application/json
      description: 将指定会话添加到免打扰列表
      parameters:
      - description: 会话ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.muteConvReq'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "409":
          description: 会话已存在
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 添加免打扰会话
      tags:
      - 设置
  /settings/mute/{convID}:
    delete:
      description: 将指定会话从免打扰列表中移除
      parameters:
      - description: 会话ID
        in: path
        name: convID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "404":
          description: 会话未找到
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 移除免打扰会话
      tags:
      - 设置
  /upload/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户头像图片，支持 jpg/png/gif/webp 格式
      parameters:
      - description: 头像图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.uploadAvatarResponse'
              type: object
        "400":
          description: 文件格式不支持或文件过大
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 上传头像
      tags:
      - 文件
securityDefinitions:
  BearerAuth:
    description: 在 Authorization 头中携带 JWT 访问令牌，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"